Refactor {card|flat} fragments, move common code to base class and fixes NPE https://github.com/VinylMusicPlayer/VinylMusicPlayer/issues/845
