+++ app/src/main/java/com/poupa/vinylmusicplayer/ui/fragments/player/PlayerAlbumCoverFragment.java
@@ -32,7 +32,7 @@ import com.poupa.vinylmusicplayer.util.ViewUtil;
  */
 public class PlayerAlbumCoverFragment extends AbsMusicServiceFragment implements ViewPager.OnPageChangeListener, MusicProgressViewUpdateHelper.Callback {
 
-    public static final int VISIBILITY_ANIM_DURATION = 300;
+    public static final long VISIBILITY_ANIM_DURATION = 300L;
 
     ViewPager viewPager;
     ImageView favoriteIcon;
