+++ app/src/main/java/com/poupa/vinylmusicplayer/ui/fragments/player/card/CardPlayerFragment.java
@@ -2,24 +2,16 @@ package com.poupa.vinylmusicplayer.ui.fragments.player.card;
 
 import android.animation.Animator;
 import android.animation.AnimatorSet;
-import android.app.Activity;
-import android.content.Context;
-import android.graphics.Color;
 import android.graphics.PorterDuff;
 import android.graphics.Typeface;
-import android.graphics.drawable.Drawable;
-import android.os.AsyncTask;
 import android.os.Build;
 import android.os.Bundle;
-import android.text.TextUtils;
 import android.view.LayoutInflater;
-import android.view.Menu;
 import android.view.MenuItem;
 import android.view.View;
 import android.view.ViewAnimationUtils;
 import android.view.ViewGroup;
 import android.view.ViewTreeObserver;
-import android.widget.FrameLayout;
 import android.widget.ImageView;
 import android.widget.TextView;
 
@@ -32,22 +24,17 @@ import androidx.recyclerview.widget.RecyclerView;
 import com.kabouzeid.appthemehelper.ThemeStore;
 import com.kabouzeid.appthemehelper.util.ATHUtil;
 import com.kabouzeid.appthemehelper.util.ColorUtil;
-import com.kabouzeid.appthemehelper.util.ToolbarContentTintHelper;
 import com.poupa.vinylmusicplayer.R;
 import com.poupa.vinylmusicplayer.adapter.base.MediaEntryViewHolder;
 import com.poupa.vinylmusicplayer.databinding.FragmentCardPlayerBinding;
 import com.poupa.vinylmusicplayer.databinding.ItemListBinding;
-import com.poupa.vinylmusicplayer.dialogs.LyricsDialog;
 import com.poupa.vinylmusicplayer.dialogs.SongShareDialog;
 import com.poupa.vinylmusicplayer.helper.MusicPlayerRemote;
 import com.poupa.vinylmusicplayer.helper.menu.SongMenuHelper;
 import com.poupa.vinylmusicplayer.misc.queue.IndexedSong;
 import com.poupa.vinylmusicplayer.model.Song;
-import com.poupa.vinylmusicplayer.model.lyrics.Lyrics;
 import com.poupa.vinylmusicplayer.ui.activities.base.AbsSlidingMusicPanelActivity;
 import com.poupa.vinylmusicplayer.ui.fragments.player.AbsPlayerFragment;
-import com.poupa.vinylmusicplayer.ui.fragments.player.PlayerAlbumCoverFragment;
-import com.poupa.vinylmusicplayer.util.ImageUtil;
 import com.poupa.vinylmusicplayer.util.MusicUtil;
 import com.poupa.vinylmusicplayer.util.PlayingSongDecorationUtil;
 import com.poupa.vinylmusicplayer.util.PreferenceUtil;
@@ -57,25 +44,18 @@ import com.poupa.vinylmusicplayer.util.VinylMusicPlayerColorUtil;
 import com.poupa.vinylmusicplayer.views.WidthFitSquareLayout;
 import com.sothree.slidinguppanel.SlidingUpPanelLayout;
 
-public class CardPlayerFragment extends AbsPlayerFragment implements PlayerAlbumCoverFragment.Callbacks, SlidingUpPanelLayout.PanelSlideListener {
-    FrameLayout toolbarContainer;
-    SlidingUpPanelLayout slidingUpPanelLayout;
-    RecyclerView recyclerView;
-    CardView playingQueueCard;
-    View colorBackground;
-    TextView playerQueueSubHeader;
+public class CardPlayerFragment extends AbsPlayerFragment implements SlidingUpPanelLayout.PanelSlideListener {
+    private SlidingUpPanelLayout slidingUpPanelLayout;
+    private RecyclerView recyclerView;
+    private CardView playingQueueCard;
+    private View colorBackground;
+    private TextView playerQueueSubHeader;
 
-    int lastColor;
+    private int lastColor;
 
-    CardPlayerPlaybackControlsFragment playbackControlsFragment;
-    PlayerAlbumCoverFragment playerAlbumCoverFragment;
+    private CardPlayerPlaybackControlsFragment playbackControlsFragment;
 
-    private AsyncTask<Song, Void, Boolean> updateIsFavoriteTask;
-    private AsyncTask<Void, Void, Lyrics> updateLyricsAsyncTask;
-
-    Lyrics lyrics;
-
-    Impl impl;
+    private Impl impl;
 
     @Override
     public View onCreateView(@NonNull LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
@@ -208,12 +188,10 @@ public class CardPlayerFragment extends AbsPlayerFragment implements PlayerAlbum
         recyclerView.getAdapter().notifyItemChanged(MusicPlayerRemote.getPosition());
     }
 
-    private void setUpSubFragments() {
+    @Override
+    protected void setUpSubFragments() {
         playbackControlsFragment = (CardPlayerPlaybackControlsFragment) getChildFragmentManager().findFragmentById(R.id.playback_controls_fragment);
-        playerAlbumCoverFragment = (PlayerAlbumCoverFragment) getChildFragmentManager().findFragmentById(R.id.player_album_cover_fragment);
-
-        if (playerAlbumCoverFragment == null) {throw new AssertionError("No fragment with id=" + R.id.player_album_cover_fragment);}
-        playerAlbumCoverFragment.setCallbacks(this);
+        super.setUpSubFragments();
     }
 
     protected void setUpPlayerToolbar() {
@@ -226,101 +204,6 @@ public class CardPlayerFragment extends AbsPlayerFragment implements PlayerAlbum
     }
 
     @Override
-    public boolean onMenuItemClick(@NonNull MenuItem item) {
-        if (item.getItemId() == R.id.action_show_lyrics) {
-            if (lyrics != null)
-                LyricsDialog.create(lyrics).show(getParentFragmentManager(), "LYRICS");
-            return true;
-        }
-        return super.onMenuItemClick(item);
-    }
-
-    private void updateIsFavorite() {
-        if (updateIsFavoriteTask != null) updateIsFavoriteTask.cancel(false);
-        updateIsFavoriteTask = new AsyncTask<Song, Void, Boolean>() {
-            @Override
-            protected Boolean doInBackground(Song... params) {
-                Activity activity = getActivity();
-                if (activity != null) {
-                    return MusicUtil.isFavorite(getActivity(), params[0]);
-                } else {
-                    cancel(false);
-                    return null;
-                }
-            }
-
-            @Override
-            protected void onPostExecute(Boolean isFavorite) {
-                Activity activity = getActivity();
-                if (activity != null) {
-                    int res = isFavorite ? R.drawable.ic_favorite_white_24dp : R.drawable.ic_favorite_border_white_24dp;
-                    int color = ToolbarContentTintHelper.toolbarContentColor(activity, Color.TRANSPARENT);
-                    Drawable drawable = ImageUtil.getTintedVectorDrawable(activity, res, color);
-                    toolbar.getMenu().findItem(R.id.action_toggle_favorite)
-                            .setIcon(drawable)
-                            .setTitle(isFavorite ? getString(R.string.action_remove_from_favorites) : getString(R.string.action_add_to_favorites));
-                }
-            }
-        }.execute(MusicPlayerRemote.getCurrentSong());
-    }
-
-    private void updateLyrics() {
-        if (updateLyricsAsyncTask != null) updateLyricsAsyncTask.cancel(false);
-        final Song song = MusicPlayerRemote.getCurrentSong();
-        if (song.equals(Song.EMPTY_SONG)) return;
-
-        updateLyricsAsyncTask = new AsyncTask<Void, Void, Lyrics>() {
-            @Override
-            protected void onPreExecute() {
-                super.onPreExecute();
-                lyrics = null;
-                playerAlbumCoverFragment.setLyrics(null);
-                toolbar.getMenu().removeItem(R.id.action_show_lyrics);
-            }
-
-            @Override
-            protected Lyrics doInBackground(Void... params) {
-                Context context = getContext();
-                if (context == null) {
-                    return null;
-                }
-                String data = MusicUtil.getLyrics(context, song);
-                if (TextUtils.isEmpty(data)) {
-                    return null;
-                }
-                return Lyrics.parse(song, data);
-            }
-
-            @Override
-            protected void onPostExecute(Lyrics l) {
-                lyrics = l;
-                playerAlbumCoverFragment.setLyrics(lyrics);
-                if (lyrics == null) {
-                    if (toolbar != null) {
-                        toolbar.getMenu().removeItem(R.id.action_show_lyrics);
-                    }
-                } else {
-                    Activity activity = getActivity();
-                    if (toolbar != null && activity != null)
-                        if (toolbar.getMenu().findItem(R.id.action_show_lyrics) == null) {
-                            int color = ToolbarContentTintHelper.toolbarContentColor(activity, Color.TRANSPARENT);
-                            Drawable drawable = ImageUtil.getTintedVectorDrawable(activity, R.drawable.ic_comment_text_outline_white_24dp, color);
-                            toolbar.getMenu()
-                                    .add(Menu.NONE, R.id.action_show_lyrics, Menu.NONE, R.string.action_show_lyrics)
-                                    .setIcon(drawable)
-                                    .setShowAsAction(MenuItem.SHOW_AS_ACTION_IF_ROOM);
-                        }
-                }
-            }
-
-            @Override
-            protected void onCancelled(Lyrics s) {
-                onPostExecute(null);
-            }
-        }.execute();
-    }
-
-    @Override
     @ColorInt
     public int getPaletteColor() {
         return lastColor;
@@ -329,17 +212,6 @@ public class CardPlayerFragment extends AbsPlayerFragment implements PlayerAlbum
     private void animateColorChange(final int newColor) {
         impl.animateColorChange(newColor);
         lastColor = newColor;
-    }
-
-    @Override
-    protected void toggleFavorite(Song song) {
-        super.toggleFavorite(song);
-        if (song.id == MusicPlayerRemote.getCurrentSong().id) {
-            if (MusicUtil.isFavorite(requireActivity(), song)) {
-                playerAlbumCoverFragment.showHeartAnimation();
-            }
-            updateIsFavorite();
-        }
     }
 
     @Override
@@ -368,29 +240,25 @@ public class CardPlayerFragment extends AbsPlayerFragment implements PlayerAlbum
     public void onColorChanged(int color) {
         animateColorChange(color);
         playbackControlsFragment.setDark(ColorUtil.isColorLight(color));
-        getCallbacks().onPaletteColorChanged();
+
+        super.onColorChanged(color);
     }
 
     @Override
-    public void onToolbarToggled() {
-        toggleToolbar(toolbarContainer);
-    }
-
-    @Override
-    public void onPanelSlide(View view, float slide) {
+    public void onPanelSlide(View panel, float slideOffset) {
         if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP) {
             float density = getResources().getDisplayMetrics().density;
-            float cardElevation = (6 * slide + 2) * density;
+            float cardElevation = (6 * slideOffset + 2) * density;
             if (isNotValidElevation(cardElevation)) return; // we have received some crash reports in setCardElevation()
             playingQueueCard.setCardElevation(cardElevation);
 
-            float buttonElevation = (2 * Math.max(0, (1 - (slide * 16))) + 2) * density;
+            float buttonElevation = (2 * Math.max(0, (1 - (slideOffset * 16))) + 2) * density;
             if (isNotValidElevation(buttonElevation)) return;
             playbackControlsFragment.playPauseFab.setElevation(buttonElevation);
         }
     }
 
-    private boolean isNotValidElevation(float elevation) {
+    private static boolean isNotValidElevation(float elevation) {
         return !(elevation >= -Float.MAX_VALUE) || !(elevation <= Float.MAX_VALUE);
     }
 
@@ -425,7 +293,7 @@ public class CardPlayerFragment extends AbsPlayerFragment implements PlayerAlbum
         void setUpPanelAndAlbumCoverHeight();
     }
 
-    private static abstract class BaseImpl implements Impl {
+    private abstract static class BaseImpl implements Impl {
         protected CardPlayerFragment fragment;
 
         public BaseImpl(CardPlayerFragment fragment) {
